refactor: removed remove method from user servise, controller & email service method renaming & separated example controller and user controller
