+++ backend-spring-boot/src/main/java/com/pharmacy/cpis/userservice/service/IPersonService.java
@@ -8,7 +8,7 @@ public interface IPersonService {
     Person findOne(Long id);
 
     List<Person> findAll();
-
+    
     Person save(Person person);
 
     void remove(Long id);
